Closes #197 enrichment tool event links are broken
Removed SaveQueryService and moved it's logic to QueryService
