+++ src/app/enrichment-table/enrichment-table-row.component.html
@@ -1,9 +1,9 @@
 <td>{{ label }}</td>
 <td [style.background-color]="getBackgroundColor(effectResult.all)">
-  <a target="_blank" [routerLink]="['/datasets', effectResult.all.countFilter.datasetId, 'browser', getStateObject(effectResult.all.countFilter) ]">{{ effectResult.all.count | numberWithExp: "1.0-2"  }}</a>
+  <a href="javascript:void(0)" (click)="goToQuery(effectResult.all.countFilter)">{{ effectResult.all.count | numberWithExp: "1.0-2"  }}</a>
 </td>
 <td [style.background-color]="getBackgroundColor(effectResult.all)">
-  <a target="_blank" [routerLink]="['/datasets', effectResult.all.overlapFilter.datasetId, 'browser', getStateObject(effectResult.all.overlapFilter) ]">{{ effectResult.all.overlapped | numberWithExp: "1.0-2"  }}</a>
+  <a href="javascript:void(0)" (click)="goToQuery(effectResult.all.overlapFilter)">{{ effectResult.all.overlapped | numberWithExp: "1.0-2"  }}</a>
 </td>
 <td [style.background-color]="getBackgroundColor(effectResult.all)">
   {{ effectResult.all.expected | numberWithExp: "1.2-2" }}
@@ -12,10 +12,10 @@
   {{ effectResult.all.pvalue | numberWithExp: "1.2-4"  }}
 </td>
 <td [style.background-color]="getBackgroundColor(effectResult.rec)">
-  <a target="_blank" [routerLink]="['/datasets', effectResult.rec.countFilter.datasetId, 'browser', getStateObject(effectResult.rec.countFilter) ]">{{ effectResult.rec.count | numberWithExp: "1.0-2"  }}</a>
+  <a href="javascript:void(0)" (click)="goToQuery(effectResult.rec.countFilter)">{{ effectResult.rec.count | numberWithExp: "1.0-2"  }}</a>
 </td>
 <td [style.background-color]="getBackgroundColor(effectResult.rec)">
-  <a target="_blank" [routerLink]="['/datasets', effectResult.rec.overlapFilter.datasetId, 'browser', getStateObject(effectResult.rec.overlapFilter) ]">{{ effectResult.rec.overlapped | numberWithExp: "1.0-2" }}</a>
+  <a href="javascript:void(0)" (click)="goToQuery(effectResult.rec.overlapFilter)">{{ effectResult.rec.overlapped | numberWithExp: "1.0-2" }}</a>
 </td>
 <td [style.background-color]="getBackgroundColor(effectResult.rec)">
   {{ effectResult.rec.expected  | numberWithExp: "1.2-2"}}
@@ -24,10 +24,10 @@
   {{ effectResult.rec.pvalue | numberWithExp: "1.2-4"}}
 </td>
 <td [style.background-color]="getBackgroundColor(effectResult.male)">
-  <a target="_blank" [routerLink]="['/datasets', effectResult.male.countFilter.datasetId, 'browser', getStateObject(effectResult.male.countFilter) ]">{{ effectResult.male.count | numberWithExp: "1.0-2"  }}</a>
+  <a href="javascript:void(0)" (click)="goToQuery(effectResult.male.countFilter)">{{ effectResult.male.count | numberWithExp: "1.0-2"  }}</a>
 </td>
 <td [style.background-color]="getBackgroundColor(effectResult.male)">
-  <a target="_blank" [routerLink]="['/datasets', effectResult.male.overlapFilter.datasetId, 'browser', getStateObject(effectResult.male.overlapFilter) ]">{{ effectResult.male.overlapped | numberWithExp: "1.0-2" }}</a>
+  <a href="javascript:void(0)" (click)="goToQuery(effectResult.male.overlapFilter)">{{ effectResult.male.overlapped | numberWithExp: "1.0-2" }}</a>
 </td>
 <td [style.background-color]="getBackgroundColor(effectResult.male)">
   {{ effectResult.male.expected | numberWithExp: "1.2-2" }}
@@ -36,10 +36,10 @@
   {{ effectResult.male.pvalue | numberWithExp: "1.2-4"}}
 </td>
 <td [style.background-color]="getBackgroundColor(effectResult.female)">
-  <a target="_blank" [routerLink]="['/datasets', effectResult.female.countFilter.datasetId, 'browser', getStateObject(effectResult.female.countFilter) ]">{{ effectResult.female.count | numberWithExp: "1.0-2"  }}</a>
+  <a href="javascript:void(0)" (click)="goToQuery(effectResult.female.countFilter)">{{ effectResult.female.count | numberWithExp: "1.0-2"  }}</a>
 </td>
 <td [style.background-color]="getBackgroundColor(effectResult.female)">
-  <a target="_blank" [routerLink]="['/datasets', effectResult.female.overlapFilter.datasetId, 'browser', getStateObject(effectResult.female.overlapFilter) ]">{{ effectResult.female.overlapped | numberWithExp: "1.0-2" }}</a>
+  <a href="javascript:void(0)" (click)="goToQuery(effectResult.female.overlapFilter)">{{ effectResult.female.overlapped | numberWithExp: "1.0-2" }}</a>
 </td>
 <td [style.background-color]="getBackgroundColor(effectResult.female)">
   {{ effectResult.female.expected  | numberWithExp: "1.2-2"}}
