+++ src/app/enrichment-table/enrichment-table-row.component.ts
@@ -4,6 +4,7 @@ import {
   GenotypePreviewWithDatasetId
 } from '../enrichment-query/enrichment-result';
 import { PValueIntensityPipe } from '../utils/p-value-intensity.pipe';
+import { QueryService } from '../query/query.service';
 
 @Component({
   selector: '[gpf-enrichment-table-row]',
@@ -13,21 +14,25 @@ import { PValueIntensityPipe } from '../utils/p-value-intensity.pipe';
 export class EnrichmentTableRowComponent {
   @Input() label: string;
   @Input() effectResult: EnrichmentEffectResult;
-  withoutDatasetId = GenotypePreviewWithDatasetId.withoutDatasetId;
-
-  getStateObject(genotypePreview: GenotypePreviewWithDatasetId): object {
-    return {
-      state: this.toJson(GenotypePreviewWithDatasetId.withoutDatasetId(genotypePreview))
-    };
-  }
-
-  toJson(obj: any): string {
-    return JSON.stringify(obj);
-  };
 
   constructor(
-    private pValueIntensityPipe: PValueIntensityPipe
+    private pValueIntensityPipe: PValueIntensityPipe,
+    private queryService: QueryService
   ) {}
+
+  goToQuery(genotypePreview: GenotypePreviewWithDatasetId) {
+    // Create new window now because we are in a 'click' event callback, update 
+    // it's url later. Otherwise this window.open is treated as a pop-up and 
+    // being blocked by most browsers.
+    // https://stackoverflow.com/a/22470171/2316754
+    let newWindow = window.open("", "_blank");
+    this.queryService.saveQuery(genotypePreview, 'genotype')
+      .take(1)
+      .subscribe(urlObject => {
+        let url = this.queryService.getLoadUrlFromResponse(urlObject);
+        newWindow.location.assign(url);
+      });
+  }
 
   getBackgroundColor(testResult: EnrichmentTestResult) {
     let intensity = this.pValueIntensityPipe.transform(testResult.pvalue);
